+++ doc/source/reference/news.rst
@@ -30,6 +30,7 @@ New features
 API Changes
 -----------
  - pagerank() and hits() now return dictionaries keyed by node
+ - clustering() and triangles() now return dictionaries keyed by node
  - add_nodes_from() now accepts (node,attrdict) two-tuples
  - eccentricity() now returns a dictionary keyed by node